listView database conditions are updated

+++ src/main/java/org/example/projectjavafx/UserProfileController.java
@@ -7,7 +7,6 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.image.ImageView;
-import javafx.scene.shape.Circle;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.text.Text;
 import javafx.event.ActionEvent;
@@ -185,34 +184,87 @@ public class UserProfileController implements Initializable {
     {
         DatabaseConnection databaseConnection = new DatabaseConnection();
         Connection connection = databaseConnection.getConnection();
-        try ( PreparedStatement preparedStatement = connection.prepareStatement("INSERT INTO friends (name) VALUES (?)"))
+        if ( !isExistInDataBase( name))
         {
-            preparedStatement.setString(1, name);
-            preparedStatement.executeUpdate();
-            addFriendList.getItems().add(name);
-            statusOfAddFriend.setText( name + " is added.");
+            statusOfAddFriend.setText( name + " not found");
         }
-        catch ( Exception e)
+        else if ( usernameLabel.getText().equals(name))
         {
-            e.printStackTrace();
+            statusOfAddFriend.setText("you cannot add yourself");
         }
+        else
+        {
+            try ( PreparedStatement preparedStatement = connection.prepareStatement("INSERT INTO friends (name) VALUES (?)"))
+            {
+                preparedStatement.setString(1, name);
+                preparedStatement.executeUpdate();
+                addFriendList.getItems().add(name);
+                statusOfAddFriend.setText( name + " is added.");
+            }
+            catch ( Exception e)
+            {
+                e.printStackTrace();
+            }
+        }
+    }
+
+    private boolean isFriendExistInTheList(String name)
+    {
+        for ( String theName : addFriendList.getItems())
+        {
+            if ( theName.equals(name))
+            {
+                return true;
+            }
+        }
+        return false;
     }
 
     private void removeFriends( String name)
     {
         DatabaseConnection databaseConnection = new DatabaseConnection();
         Connection connection = databaseConnection.getConnection();
-        try ( PreparedStatement preparedStatement = connection.prepareStatement("DELETE FROM friends WHERE name = ?"))
+        if ( isFriendExistInTheList( name))
         {
-            preparedStatement.setString(1, name);
-            preparedStatement.executeUpdate();
-            addFriendList.getItems().remove(name);
-            statusOfAddFriend.setText( name + " is removed.");
+            try ( PreparedStatement preparedStatement = connection.prepareStatement("DELETE FROM friends WHERE name = ?"))
+            {
+                preparedStatement.setString(1, name);
+                preparedStatement.executeUpdate();
+                addFriendList.getItems().remove(name);
+                statusOfAddFriend.setText( name + " is removed.");
+            }
+            catch (Exception e)
+            {
+                e.printStackTrace();
+            }
+        }
+        else
+        {
+            statusOfAddFriend.setText( name + " is not your friend.");
+        }
+    }
+
+    private boolean isExistInDataBase( String name)
+    {
+        DatabaseConnection databaseConnection = new DatabaseConnection();
+        Connection connection = databaseConnection.getConnection();
+
+        String sql = "SELECT COUNT(*) FROM friends WHERE name = ?";
+        try(PreparedStatement statement = connection.prepareStatement(sql))
+        {
+            statement.setString(1, name);
+            ResultSet resultSet = statement.executeQuery();
+            if ( resultSet.next())
+            {
+                return resultSet.getInt(1) > 0;
+            }
+
         }
         catch (Exception e)
         {
             e.printStackTrace();
         }
+        return false;
     }
 
     public void changePassword() {
